requirements.txt
streamlit==1.31.1
pymupdf==1.23.26
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
reportlab